refactor: rename theme.dart -> text_theme.dart as it now only contains the text theme.
