+++ client/lib/styles/styles.dart
@@ -1,4 +1,4 @@
-export 'theme.dart';
+export 'text_theme.dart';
 export 'theme_util.dart';
 export 'app_styles.dart';
 export 'material_theme.dart';
